Implementimi i StrategyPattern

+++ src/main/java/com/example/demo/repository/JsPytjeRepository.java
@@ -15,7 +15,7 @@ public interface JsPytjeRepository  extends JpaRepository<JsPytjeEntity,Long> {
     public List<String> findByPergjigja();
 
     @Query(value = "select * from js_pytje where KuiziID = ?1",nativeQuery = true)
-    public List<JsPytjeEntity> findAllByQuizID(Integer id);
+    public List<JsPytjeEntity> findAllByQuizID(Long id);
 
 
 }
